+++ src/ui/preferences/general/mod.rs
@@ -1,29 +1,19 @@
 use relm4::prelude::*;
-
-use relm4::factory::{
-    AsyncFactoryVecDeque,
-    AsyncFactoryComponent,
-    AsyncFactorySender
-};
-
+use relm4::factory::{AsyncFactoryComponent, AsyncFactorySender, AsyncFactoryVecDeque};
 use gtk::prelude::*;
 use adw::prelude::*;
-
 use anime_launcher_sdk::wincompatlib::prelude::*;
-
 use anime_launcher_sdk::anime_game_core::prelude::*;
 use anime_launcher_sdk::anime_game_core::star_rail::consts::GameEdition;
-
 use anime_launcher_sdk::config::ConfigExt;
 use anime_launcher_sdk::star_rail::config::Config;
-use anime_launcher_sdk::star_rail::config::schema::launcher::{LauncherStyle, LauncherBehavior};
+use anime_launcher_sdk::star_rail::config::schema::launcher::{LauncherBehavior, LauncherStyle};
 
 pub mod components;
 
 use components::*;
 
 use crate::ui::preferences::main::PreferencesAppMsg;
-
 use crate::i18n::*;
 use crate::*;
 
@@ -36,10 +26,10 @@ struct VoicePackageComponent {
 
 #[relm4::factory(async)]
 impl AsyncFactoryComponent for VoicePackageComponent {
+    type CommandOutput = ();
     type Init = (VoiceLocale, bool);
     type Input = GeneralAppMsg;
     type Output = GeneralAppMsg;
-    type CommandOutput = ();
     type ParentWidget = adw::ExpanderRow;
 
     view! {
@@ -83,7 +73,7 @@ impl AsyncFactoryComponent for VoicePackageComponent {
     async fn init_model(
         init: Self::Init,
         _index: &DynamicIndex,
-        _sender: AsyncFactorySender<Self>,
+        _sender: AsyncFactorySender<Self>
     ) -> Self {
         Self {
             locale: init.0,
@@ -95,8 +85,7 @@ impl AsyncFactoryComponent for VoicePackageComponent {
     async fn update(&mut self, msg: Self::Input, sender: AsyncFactorySender<Self>) {
         self.installed = !self.installed;
 
-        sender.output(msg)
-            .unwrap();
+        sender.output(msg).unwrap();
     }
 }
 
@@ -108,17 +97,19 @@ pub struct GeneralApp {
     main_patch: Option<(Version, JadeitePatchStatusVariant)>,
 
     style: LauncherStyle,
+    use_video_background: bool,
+    background_index: u8,
     languages: Vec<String>
 }
 
 #[derive(Debug, Clone)]
 pub enum GeneralAppMsg {
-    /// Supposed to be called automatically on app's run when the latest game version
-    /// was retrieved from the API
+    /// Supposed to be called automatically on app's run when the latest game
+    /// version was retrieved from the API
     SetGameDiff(Option<VersionDiff>),
 
-    /// Supposed to be called automatically on app's run when the latest UnityPlayer patch version
-    /// was retrieved from remote repos
+    /// Supposed to be called automatically on app's run when the latest
+    /// UnityPlayer patch version was retrieved from remote repos
     SetMainPatch(Option<(Version, JadeitePatchStatusVariant)>),
 
     // If one ever wish to change it to accept VoiceLocale
@@ -138,6 +129,7 @@ pub enum GeneralAppMsg {
     OpenComponentsPage,
 
     UpdateLauncherStyle(LauncherStyle),
+    SetVideoBackground(bool),
 
     WineOpen(&'static [&'static str]),
 
@@ -243,6 +235,20 @@ impl SimpleAsyncComponent for GeneralApp {
                             } else {
                                 std::fs::write(KEEP_BACKGROUND_FILE.as_path(), "");
                             }
+                        }
+                    }
+                },
+
+                adw::ActionRow {
+                    set_title: &tr!("video-background"),
+                    set_subtitle: &tr!("video-background-description"),
+
+                    add_suffix = &gtk::Switch {
+                        set_valign: gtk::Align::Center,
+                        set_active: model.use_video_background,
+
+                        connect_state_notify[sender] => move |switch| {
+                            sender.input(GeneralAppMsg::SetVideoBackground(switch.state()));
                         }
                     }
                 }
@@ -549,7 +555,7 @@ impl SimpleAsyncComponent for GeneralApp {
     async fn init(
         _init: Self::Init,
         root: Self::Root,
-        sender: AsyncComponentSender<Self>,
+        sender: AsyncComponentSender<Self>
     ) -> AsyncComponentParts<Self> {
         tracing::info!("Initializing general settings");
 
@@ -566,13 +572,22 @@ impl SimpleAsyncComponent for GeneralApp {
             main_patch: None,
 
             style: CONFIG.launcher.style,
-            languages: SUPPORTED_LANGUAGES.iter().map(|lang| tr!(format_lang(lang).as_str())).collect()
+            use_video_background: CONFIG.launcher.video_background,
+            background_index: CONFIG.launcher.background_index,
+            languages: SUPPORTED_LANGUAGES
+                .iter()
+                .map(|lang| tr!(format_lang(lang).as_str()))
+                .collect()
         };
 
         for package in VoiceLocale::list() {
             model.voice_packages.guard().push_back((
                 *package,
-                CONFIG.game.voices.iter().any(|voice| VoiceLocale::from_str(voice) == Some(*package))
+                CONFIG
+                    .game
+                    .voices
+                    .iter()
+                    .any(|voice| VoiceLocale::from_str(voice) == Some(*package))
             ));
         }
 
@@ -581,7 +596,10 @@ impl SimpleAsyncComponent for GeneralApp {
 
         let widgets = view_output!();
 
-        AsyncComponentParts { model, widgets }
+        AsyncComponentParts {
+            model,
+            widgets
+        }
     }
 
     async fn update(&mut self, msg: Self::Input, sender: AsyncComponentSender<Self>) {
@@ -600,8 +618,16 @@ impl SimpleAsyncComponent for GeneralApp {
             GeneralAppMsg::AddVoicePackage(index) => {
                 if let Some(package) = self.voice_packages.get(index.current_index()) {
                     if let Ok(mut config) = Config::get() {
-                        if !config.game.voices.iter().any(|voice| VoiceLocale::from_str(voice) == Some(package.locale)) {
-                            config.game.voices.push(package.locale.to_code().to_string());
+                        if !config
+                            .game
+                            .voices
+                            .iter()
+                            .any(|voice| VoiceLocale::from_str(voice) == Some(package.locale))
+                        {
+                            config
+                                .game
+                                .voices
+                                .push(package.locale.to_code().to_string());
 
                             Config::update(config);
 
@@ -617,17 +643,29 @@ impl SimpleAsyncComponent for GeneralApp {
                     if let Ok(mut config) = Config::get() {
                         package.sensitive = false;
 
-                        config.game.voices.retain(|voice| VoiceLocale::from_str(voice) != Some(package.locale));
+                        config
+                            .game
+                            .voices
+                            .retain(|voice| VoiceLocale::from_str(voice) != Some(package.locale));
 
                         Config::update(config.clone());
 
-                        let package = VoicePackage::with_locale(package.locale, config.launcher.edition).unwrap();
-                        let game_path = config.game.path.for_edition(config.launcher.edition).to_path_buf();
+                        let package =
+                            VoicePackage::with_locale(package.locale, config.launcher.edition)
+                                .unwrap();
+                        let game_path = config
+                            .game
+                            .path
+                            .for_edition(config.launcher.edition)
+                            .to_path_buf();
 
                         if package.is_installed_in(&game_path) {
                             std::thread::spawn(move || {
                                 if let Err(err) = package.delete_in(game_path) {
-                                    tracing::error!("Failed to delete voice package: {:?}", package.locale());
+                                    tracing::error!(
+                                        "Failed to delete voice package: {:?}",
+                                        package.locale()
+                                    );
 
                                     sender.input(GeneralAppMsg::Toast {
                                         title: tr!("voice-package-deletion-error"),
@@ -635,12 +673,11 @@ impl SimpleAsyncComponent for GeneralApp {
                                     });
                                 }
 
-                                sender.input(GeneralAppMsg::SetVoicePackageSensitivity(index, true));
+                                sender
+                                    .input(GeneralAppMsg::SetVoicePackageSensitivity(index, true));
                                 sender.output(PreferencesAppMsg::UpdateLauncherState);
                             });
-                        }
-
-                        else {
+                        } else {
                             sender.input(GeneralAppMsg::SetVoicePackageSensitivity(index, true));
                         }
                     }
@@ -654,13 +691,15 @@ impl SimpleAsyncComponent for GeneralApp {
             }
 
             GeneralAppMsg::UpdateDownloadedWine => {
-                self.components_page.sender()
+                self.components_page
+                    .sender()
                     .send(ComponentsPageMsg::UpdateDownloadedWine)
                     .unwrap();
             }
 
             GeneralAppMsg::UpdateDownloadedDxvk => {
-                self.components_page.sender()
+                self.components_page
+                    .sender()
                     .send(ComponentsPageMsg::UpdateDownloadedDxvk)
                     .unwrap();
             }
@@ -676,25 +715,30 @@ impl SimpleAsyncComponent for GeneralApp {
             // Don't care about it, don't want to rewrite everything.
             #[allow(static_mut_refs)]
             GeneralAppMsg::OpenMainPage => unsafe {
-                PREFERENCES_WINDOW.as_ref()
+                PREFERENCES_WINDOW
+                    .as_ref()
                     .unwrap_unchecked()
                     .widget()
                     .pop_subpage();
-            }
+            },
 
             // Don't care about it, don't want to rewrite everything.
             #[allow(static_mut_refs)]
             GeneralAppMsg::OpenComponentsPage => unsafe {
-                PREFERENCES_WINDOW.as_ref()
+                PREFERENCES_WINDOW
+                    .as_ref()
                     .unwrap_unchecked()
                     .widget()
                     .push_subpage(self.components_page.widget());
-            }
+            },
 
             GeneralAppMsg::UpdateLauncherStyle(style) => {
                 if style == LauncherStyle::Classic && !KEEP_BACKGROUND_FILE.exists() {
-                    if let Err(err) = crate::background::download_background() {
-                        tracing::error!("Failed to download background picture");
+                    if let Err(err) = crate::background::download_background(
+                        self.use_video_background,
+                        self.background_index
+                    ) {
+                        tracing::error!("Failed to download background picture/video");
 
                         sender.input(GeneralAppMsg::Toast {
                             title: tr!("background-downloading-failed"),
@@ -716,12 +760,45 @@ impl SimpleAsyncComponent for GeneralApp {
                 let _ = sender.output(Self::Output::SetLauncherStyle(style));
             }
 
+            GeneralAppMsg::SetVideoBackground(use_video) => {
+                if self.style == LauncherStyle::Classic
+                    && !KEEP_BACKGROUND_FILE.exists()
+                    && use_video
+                {
+                    if let Err(err) =
+                        crate::background::download_background(use_video, self.background_index)
+                    {
+                        tracing::error!("Failed to download background picture/video");
+
+                        sender.input(GeneralAppMsg::Toast {
+                            title: tr!("background-downloading-failed"),
+                            description: Some(err.to_string())
+                        });
+
+                        return;
+                    }
+                }
+
+                if let Ok(mut config) = Config::get() {
+                    config.launcher.video_background = use_video;
+
+                    Config::update(config);
+                }
+
+                self.use_video_background = use_video;
+
+                let _ = sender.output(Self::Output::SetVideoBackground(use_video));
+            }
+
             GeneralAppMsg::WineOpen(executable) => {
                 let config = Config::get().unwrap_or_else(|_| CONFIG.clone());
 
                 if let Ok(Some(wine)) = config.get_selected_wine() {
                     let result = wine
-                        .to_wine(config.components.path, Some(config.game.wine.builds.join(&wine.name)))
+                        .to_wine(
+                            config.components.path,
+                            Some(config.game.wine.builds.join(&wine.name))
+                        )
                         .with_prefix(config.game.wine.prefix)
                         .with_loader(WineLoader::Current)
                         .with_arch(WineArch::Win64)
@@ -729,9 +806,7 @@ impl SimpleAsyncComponent for GeneralApp {
 
                     if let Err(err) = result {
                         sender.input(GeneralAppMsg::Toast {
-                            title: tr!("wine-run-error", {
-                                "executable" = executable.join(" ")
-                            }),
+                            title: tr!("wine-run-error", { "executable" = executable.join(" ") }),
                             description: Some(err.to_string())
                         });
 
@@ -740,8 +815,14 @@ impl SimpleAsyncComponent for GeneralApp {
                 }
             }
 
-            GeneralAppMsg::Toast { title, description } => {
-                let _ = sender.output(Self::Output::Toast { title, description });
+            GeneralAppMsg::Toast {
+                title,
+                description
+            } => {
+                let _ = sender.output(Self::Output::Toast {
+                    title,
+                    description
+                });
             }
         }
     }